+++ kagra-shared/src/collectathon.rs
@@ -411,6 +411,7 @@ impl CollectathonScene {
             fog_start: 48.0,
             fog_end: 150.0,
             batches: b.finish(),
+            ..Default::default()
         }
     }